+++ ureport/polls/migrations/0024_auto_20160118_0934.py
@@ -14,6 +14,6 @@ class Migration(migrations.Migration):
         migrations.AlterField(
             model_name='poll',
             name='poll_date',
-            field=models.DateTimeField(null=True),
+            field=models.DateTimeField(),
         ),
     ]