+++ ureport/polls/views.py
@@ -1,16 +1,16 @@
 import json
 
+import six
 from dash.orgs.views import OrgPermsMixin, OrgObjPermsMixin
 from django import forms
-from django.core.files.base import ContentFile
-from django.core.urlresolvers import reverse
+from django.urls import reverse
 from dash.categories.models import Category, CategoryImage
 from dash.categories.fields import CategoryChoiceField
 from django.utils import timezone
 from smartmin.csv_imports.models import ImportTask
 
 from ureport.utils import json_date_to_datetime
-from .models import Poll, PollQuestion, FeaturedResponse, PollImage, CACHE_ORG_FLOWS_KEY
+from .models import Poll, PollQuestion, FeaturedResponse, PollImage
 from smartmin.views import SmartCRUDL, SmartCreateView, SmartListView, SmartUpdateView, SmartCSVImportView
 from django.utils.translation import ugettext_lazy as _
 from django.core.exceptions import ValidationError
@@ -152,8 +152,8 @@ class PollCRUDL(SmartCRUDL):
         title = _("Poll Images")
         success_message = _("Now enter any responses you'd like to feature. (if any)")
 
-        def get_form(self, form_class):
-            form = super(PollCRUDL.Images, self).get_form(form_class)
+        def get_form(self):
+            form = super(PollCRUDL.Images, self).get_form()
             form.fields.clear()
 
             idx = 1
@@ -196,8 +196,8 @@ class PollCRUDL(SmartCRUDL):
         title = _("Featured Poll Responses")
         success_message = _("Your poll has been updated.")
 
-        def get_form(self, form_class):
-            form = super(PollCRUDL.Responses, self).get_form(form_class)
+        def get_form(self):
+            form = super(PollCRUDL.Responses, self).get_form()
             form.fields.clear()
 
             existing_responses = list(self.object.featured_responses.all().order_by('pk'))
@@ -266,8 +266,8 @@ class PollCRUDL(SmartCRUDL):
         def get_questions(self):
             return self.object.questions.all().order_by('-priority', 'pk')
 
-        def get_form(self, form_class):
-            form = super(PollCRUDL.Questions, self).get_form(form_class)
+        def get_form(self):
+            form = super(PollCRUDL.Questions, self).get_form()
             form.fields.clear()
 
             # fetch this single flow so we load what rules are available
@@ -435,7 +435,7 @@ class PollCRUDL(SmartCRUDL):
         def post_save(self, task):
             # configure import params with current org and timezone
             org = self.request.org
-            params = dict(org_id=org.id, timezone=org.timezone, original_filename=self.form.cleaned_data['csv_file'].name)
+            params = dict(org_id=org.id, timezone=six.text_type(org.timezone), original_filename=self.form.cleaned_data['csv_file'].name)
             params_dump = json.dumps(params)
             ImportTask.objects.filter(pk=task.pk).update(import_params=params_dump)